Added argument validation for repositories

+++ WeatherForecast.DataAccess/Repositories/CityRepository.cs
@@ -1,4 +1,6 @@
-﻿using System.Collections.Generic;
+﻿using System;
+using System.Collections.Generic;
+using System.ComponentModel.DataAnnotations;
 using System.Linq;
 using System.Threading.Tasks;
 using Dapper;
@@ -11,6 +13,8 @@ namespace WeatherForecast.DataAccess.Repositories
     {
         public async Task<City> GetByIdAsync(int id)
         {
+            if (!_validator.IsValidId(id)) throw new ArgumentException("Id must be more then 0");
+
             return await _queryExecutor.ExecuteAsync(async (db) => await db.QueryFirstAsync<City>($"SELECT * FROM {TableNames.cities} WHERE id = '{id}'"));
         }
 
@@ -21,6 +25,9 @@ namespace WeatherForecast.DataAccess.Repositories
 
         public async Task<City> AddAsync(City entity)
         {
+            if (!_validator.IsNotNull(entity)) throw new NullReferenceException();
+            if (!_validator.IsValidName(entity.Name)) throw new ValidationException($"{nameof(entity.Name)} can not be empty");
+
             return await _queryExecutor.ExecuteAsync(async (db) =>
             {
                 var sqlQuery = $"INSERT INTO {TableNames.cities}(name) " +
@@ -33,6 +40,9 @@ namespace WeatherForecast.DataAccess.Repositories
 
         public async Task<City> AddOrUpdateAsync(City entity)
         {
+            if (!_validator.IsNotNull(entity)) throw new NullReferenceException();
+            if (!_validator.IsValidName(entity.Name)) throw new ValidationException($"{nameof(entity.Name)} can not be empty");
+
             return await _queryExecutor.ExecuteAsync(async (db) =>
             {
                 var sqlQuery = $"SELECT id FROM {TableNames.cities} " +
@@ -52,6 +62,10 @@ namespace WeatherForecast.DataAccess.Repositories
 
         public async Task UpdateAsync(City entity)
         {
+            if (!_validator.IsNotNull(entity)) throw new NullReferenceException();
+            if (!_validator.IsValidId(entity.Id)) throw new ArgumentException($"{nameof(entity.Id)} must be more then 0");
+            if (!_validator.IsValidName(entity.Name)) throw new ValidationException($"{nameof(entity.Name)} can not be empty");
+
             await _queryExecutor.ExecuteAsync<Task>(async (db) =>
             {
                 var sqlQuery = $"UPDATE {TableNames.cities} " +
@@ -63,6 +77,10 @@ namespace WeatherForecast.DataAccess.Repositories
 
         public async Task DeleteAsync(City entity)
         {
+            if (!_validator.IsNotNull(entity)) throw new NullReferenceException();
+            if (!_validator.IsValidId(entity.Id)) throw new ArgumentException($"{nameof(entity.Id)} must be more then 0");
+            if (!_validator.IsValidName(entity.Name)) throw new ValidationException($"{nameof(entity.Name)} can not be empty");
+
             await _queryExecutor.ExecuteAsync<Task>(async (db) =>
             {
                 var sqlQuery = $"DELETE FROM {TableNames.weather} " +
